+++ .idea/runConfigurations/Deno__main_ts.xml
@@ -1,0 +1,6 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="Deno: main.ts" type="DenoConfigurationType" inputPath="$PROJECT_DIR$/main.ts" programParameters="run --unstable --allow-net --allow-run --import-map=import_map.json">
+    <option name="applicationArguments" value="$Prompt$" />
+    <method v="2" />
+  </configuration>
+</component>
