+++ .idea/runConfigurations/Deno__main_ts_write_lock.xml
@@ -1,0 +1,6 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="Deno: main.ts write lock" type="DenoConfigurationType" inputPath="$PROJECT_DIR$/main.ts" programParameters="run --unstable --allow-net --allow-run --import-map=import_map.json --lock=lock.json --lock-write">
+    <option name="applicationArguments" value="$Prompt$" />
+    <method v="2" />
+  </configuration>
+</component>
