--- .idea/runConfigurations/Deno__mod_ts.xml
@@ -1,6 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Deno: mod.ts" type="DenoConfigurationType" nameIsGenerated="true" inputPath="$PROJECT_DIR$/mod.ts" programParameters="run --unstable --allow-net --allow-run --import-map=import_map.json">
-    <option name="applicationArguments" value="$Prompt$" />
-    <method v="2" />
-  </configuration>
-</component>
