+++ .idea/runConfigurations/Deno_test.xml
@@ -1,0 +1,5 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="Deno test" type="DenoConfigurationType" inputPath="." programParameters="test --unstable --allow-net --allow-run --import-map=import_map.json">
+    <method v="2" />
+  </configuration>
+</component>
