+++ import_map.json
@@ -4,6 +4,7 @@
     "./": "./",
     "std/": "https://deno.land/std@0.105.0/",
     "oak": "https://deno.land/x/oak@v9.0.1/mod.ts",
-    "oak_logger": "https://deno.land/x/oak_logger@1.0.0/mod.ts"
+    "oak_logger": "https://deno.land/x/oak_logger@1.0.0/mod.ts",
+    "superoak": "https://deno.land/x/superoak@4.4.0/mod.ts"
   }
 }
